requirements.txt
udao==1.0.5
pre-commit==3.4.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-mock==3.12.0
chardet==5.2.0
sphinx==7.2.6
sphinx-book-theme==1.1.0
torch_geometric==2.5.3
